Removed php-imap extension
The imap extension is not needed by Magento. It was included here simply because it was in the default list of extensions in davidalger.php_fpm ansible role, likely due to that role having been used once upon a time for deployment of a ticketing system which used imap accounts. Since this is not needed by Magento and the imap extension is not available to us in EL 8 streams, it has been removed completely. It can be added downstream later on if ever needed by a custom application.